move the product-list ui to entities
