+++ src/entities/product/ui/index.ts
@@ -1,1 +1,2 @@
 export { ProductCard } from './product-card';
+export { ProductList } from './product-list';
