--- src/widgets/catalog/ui/product-list/products/index.ts
@@ -1,1 +1,0 @@
-export { Products } from './products';
