--- src/widgets/catalog/ui/product-list/products/products.types.ts
@@ -1,5 +1,0 @@
-import type { ShortInfo } from '@/entities/product';
-
-export type ProductsProps = {
-  products: ShortInfo[];
-};
